requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.1
Pillow==10.1.0
aiohttp==3.9.1
dnspython==2.4.2
